requirements.txt
pymongo>=4.6.0
openai>=1.12.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-mongodb>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.29.0
plotly>=5.17.0

